Added the PYNQ deployment code.
In order to make deploying this demo as easy as possible it's important to keep everything in one place. I added the necessary code to deploy to a PYNQ board from @ben-hawks.

As a result only this repository needs to be cloned to the PYNQ board which will make deployment faster.